bitmanipulation: fix undefined behavior in bit shift in getbe32
OSS-Fuzz caught this in its ubsan mode[0]. I'm not worried about a
security issue here because in practice this should work out the way
we naively expected, we're just making things explicit to the
compiler with the casts.

0: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=8876

Differential Revision: https://phab.mercurial-scm.org/D3729

+++ mercurial/bitmanipulation.h
@@ -9,7 +9,8 @@ static inline uint32_t getbe32(const char *c)
 {
 	const unsigned char *d = (const unsigned char *)c;
 
-	return ((d[0] << 24) | (d[1] << 16) | (d[2] << 8) | (d[3]));
+	return ((((uint32_t)d[0]) << 24) | (((uint32_t)d[1]) << 16) |
+	        (((uint32_t)d[2]) << 8) | (d[3]));
 }
 
 static inline int16_t getbeint16(const char *c)
